Refactor: Remove unused component and improve BottomSheet
This commit removes an unused UI component and optimizes the `BottomSheet` implementation.

- **NavigationTile.kt:** Deleted the file as the component is no longer in use.
- **BottomSheet.kt:**
    - Simplified the alpha calculation logic in `graphicsLayer` by using `pow(0.5f)` directly, removing unnecessary type conversions.
    - Added a comment explaining the background fade behavior.
    - Improved code formatting and removed extra whitespace.

+++ app/src/main/java/com/arturo254/opentune/ui/component/BottomSheet.kt
@@ -48,6 +48,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import kotlin.math.pow
 
+
 /**
  * Bottom Sheet
  * Modified from [ViMusic](https://github.com/vfsfitvnm/ViMusic)
@@ -57,23 +58,19 @@ fun BottomSheet(
     state: BottomSheetState,
     modifier: Modifier = Modifier,
     background: @Composable (BoxScope.() -> Unit) = { },
-
     onDismiss: (() -> Unit)? = null,
     collapsedContent: @Composable BoxScope.() -> Unit,
     content: @Composable BoxScope.() -> Unit,
 ) {
-
-
-            Box(
-                modifier = modifier
-                    .graphicsLayer {
-                        alpha = (1.4f * ((state.progress.coerceAtLeast(0.1f) - 0.1f).toDouble().pow(0.5)).toFloat())
-                            .coerceIn(0f, 1f)
-                    }
-                    .fillMaxSize(),
-                content = background
-            )
-
+    Box(
+        modifier = modifier
+            .graphicsLayer {
+                // background fades during about 10%-61% progress
+                alpha = (1.4f * (state.progress.coerceAtLeast(0.1f) - 0.1f).pow(0.5f)).coerceIn(0f, 1f)
+            }
+            .fillMaxSize(),
+        content = background
+    )
     Box(
         modifier = modifier
             .fillMaxSize()
@@ -108,7 +105,6 @@ fun BottomSheet(
                     topEnd = if (!state.isExpanded) 16.dp else 0.dp
                 )
             )
-
     ) {
         if (!state.isCollapsed && !state.isDismissed) {
             BackHandler(onBack = state::collapseSoft)
